+++ python/cudf/cudf/_libxx/__init__.py
@@ -4,6 +4,7 @@ import numpy as np
 from . import (
     copying,
     hash,
+    join,
     merge,
     null_mask,
     quantiles,