+++ python/cudf/cudf/_libxx/aggregation.pyx
@@ -18,7 +18,7 @@ from cudf._libxx.cpp.types cimport (
     data_type,
     interpolation
 )
-cimport cudf._libxx.cpp.aggregation as cudf_aggregation
+cimport cudf._libxx.cpp.aggregation as libcudf_aggregation
 
 
 # need to update as and when we add new aggregations with additional options
@@ -30,15 +30,15 @@ cdef unique_ptr[aggregation] get_aggregation(op, kwargs) except *:
     cdef unique_ptr[aggregation] agg
 
     if op == "sum":
-        agg = move(cudf_aggregation.make_sum_aggregation())
+        agg = move(libcudf_aggregation.make_sum_aggregation())
     elif op == "min":
-        agg = move(cudf_aggregation.make_min_aggregation())
+        agg = move(libcudf_aggregation.make_min_aggregation())
     elif op == "max":
-        agg = move(cudf_aggregation.make_max_aggregation())
+        agg = move(libcudf_aggregation.make_max_aggregation())
     elif op == "mean":
-        agg = move(cudf_aggregation.make_mean_aggregation())
+        agg = move(libcudf_aggregation.make_mean_aggregation())
     elif op == "count":
-        agg = move(cudf_aggregation.make_count_aggregation())
+        agg = move(libcudf_aggregation.make_count_aggregation())
     elif callable(op):
         # Handling UDF type
         nb_type = numba.numpy_support.from_dtype(kwargs['dtype'])
@@ -55,8 +55,8 @@ cdef unique_ptr[aggregation] get_aggregation(op, kwargs) except *:
 
         out_dtype = data_type(tid)
 
-        agg = move(cudf_aggregation.make_udf_aggregation(
-            cudf_aggregation.udf_type.PTX, cpp_str, out_dtype
+        agg = move(libcudf_aggregation.make_udf_aggregation(
+            libcudf_aggregation.udf_type.PTX, cpp_str, out_dtype
         ))
     else:
         assert False, "Invalid aggreagtion operation"